QH/QHSExo raws: strip double() wrapper on reshape(maxindex,...) linidx
- Drop `double(...)` around `reshape(maxindexL2,[...])` / `reshape(maxindex,[...])`
  index build across QH/QHSExo VFI + fastOLG/slowOLG TPath GI raws (64 files);
  remove leftover commented `double(Phi_aprimeMatrix)` and strip `double()`
  inside commented surf/plot debug block in EntryExit subfn.
- ExpAsset fastOLG DC1 (e and noe): flip Refine flag to 1 in
  CreateReturnFnMatrix_fastOLG_ExpAsset_Disc[_e]; reshape DiscountedEV
  for broadcast against [N_d,N_a1,level1n,N_a2,N_j,N_z(,N_e)].
- aprimeFnMatrix experience-asset l_a2==2: replace 4-corner Kron-folded
  output with per-dim factored [l_a2,...] indexes/probs (caller does
  nested 2-corner interp w/ skipinterp); add paramShift for
  fastOLG==0 && N_semizze>0 branch in CreateaprimePolicyExperienceAsset_J.

+++ TransitionPaths/FHorz/subcodes/ValueFnSingleStep/QuasiHyperbolic/fastOLG/DivideConquer/GridInterpLayer/ValueFnIter_FHorz_TPath_SingleStep_QHS_fastOLG_DC1_GI1_e_raw.m
@@ -129,7 +129,7 @@ if vfoptions.lowmemory==0
     Policy(4,:,:,:,:)=shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
     % Vunderbar = Vhat + (beta - beta0beta)*EV_at_policy
-    linidx=double(reshape(maxindexL2,[1,N_a*N_j*N_z*N_e]))+N_d*n2long*(0:N_a*N_j*N_z*N_e-1);
+    linidx=reshape(maxindexL2,[1,N_a*N_j*N_z*N_e])+N_d*n2long*(0:N_a*N_j*N_z*N_e-1);
     EV_at_policy=reshape(EVfine(linidx),[N_a*N_j,N_z,N_e]);
     Vhat=reshape(Vhat,[N_a*N_j,N_z,N_e]);
     V=Vhat+EV_at_policy;
@@ -190,7 +190,7 @@ elseif vfoptions.lowmemory==1
         inUpperStrict=(L2offset>=n2short+3) & (L2offset<=n2long-1);
         Policy(4,:,:,:,e_c)=shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
-        linidx_e=double(reshape(maxindexL2,[1,N_a*N_j*N_z]))+N_d*n2long*(0:N_a*N_j*N_z-1);
+        linidx_e=reshape(maxindexL2,[1,N_a*N_j*N_z])+N_d*n2long*(0:N_a*N_j*N_z-1);
         EV_at_policy_e=reshape(EVfine_e(linidx_e),[N_a*N_j,N_z]);
         Vhat(:,:,e_c)=reshape(Vhat_e,[N_a*N_j,N_z]);
         V(:,:,e_c)=Vhat(:,:,e_c)+EV_at_policy_e;
@@ -259,7 +259,7 @@ elseif vfoptions.lowmemory==2
             inUpperStrict=(L2offset>=n2short+3) & (L2offset<=n2long-1);
             Policy(4,:,:,z_c,e_c)=shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
-            linidx_ze=double(reshape(maxindexL2,[1,N_a*N_j]))+N_d*n2long*(0:N_a*N_j-1);
+            linidx_ze=reshape(maxindexL2,[1,N_a*N_j])+N_d*n2long*(0:N_a*N_j-1);
             EV_at_policy_ze=reshape(EVfine_ze(linidx_ze),[N_a*N_j,1]);
             Vhat(:,z_c,e_c)=reshape(Vhat_ze,[N_a*N_j,1]);
             V(:,z_c,e_c)=Vhat(:,z_c,e_c)+EV_at_policy_ze;
