+++ TransitionPaths/FHorz/subcodes/ValueFnSingleStep/QuasiHyperbolic/fastOLG/GridInterpLayer/ValueFnIter_FHorz_TPath_SingleStep_QHS_fastOLG_GI1_nod_e_raw.m
@@ -72,7 +72,7 @@ if vfoptions.lowmemory==0
     Policy(3,:,:,:,:) = shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
     %% Vunderbar: re-evaluate at Policy's aprime with beta
-    linidx=double(reshape(maxindexL2,[1,N_a*N_j*N_z*N_e]))+n2long*(0:N_a*N_j*N_z*N_e-1);
+    linidx=reshape(maxindexL2,[1,N_a*N_j*N_z*N_e])+n2long*(0:N_a*N_j*N_z*N_e-1);
     EV_at_policy=reshape(EVfine(linidx),[N_a,N_j,N_z,N_e]);
     V=shiftdim(Vhatii,1)+reshape(beta_J-beta0beta_J,[1,N_j,1,1]).*EV_at_policy;
     V=reshape(V,[N_a*N_j,N_z,N_e]);
@@ -109,7 +109,7 @@ elseif vfoptions.lowmemory==1
         Policy(3,:,:,:,e_c) = shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
         %% Vunderbar
-        linidx_e=double(reshape(maxindexL2,[1,N_a*N_j*N_z]))+n2long*(0:N_a*N_j*N_z-1);
+        linidx_e=reshape(maxindexL2,[1,N_a*N_j*N_z])+n2long*(0:N_a*N_j*N_z-1);
         EV_at_policy_e=reshape(EVfine_e(linidx_e),[N_a,N_j,N_z]);
         Vtemp=shiftdim(Vhatii,1)+reshape(beta_J-beta0beta_J,[1,N_j,1]).*EV_at_policy_e;
         V(:,:,e_c)=reshape(Vtemp,[N_a*N_j,N_z]);
@@ -154,7 +154,7 @@ elseif vfoptions.lowmemory==2
             Policy(3,:,:,z_c,e_c) = shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
             %% Vunderbar
-            linidx_ze=double(reshape(maxindexL2,[1,N_a*N_j]))+n2long*(0:N_a*N_j-1);
+            linidx_ze=reshape(maxindexL2,[1,N_a*N_j])+n2long*(0:N_a*N_j-1);
             EV_at_policy_ze=reshape(EVfine_ze(linidx_ze),[N_a,N_j]);
             Vtemp=shiftdim(Vhatii,1)+reshape(beta_J-beta0beta_J,[1,N_j]).*EV_at_policy_ze;
             V(:,z_c,e_c)=reshape(Vtemp,[N_a*N_j,1]);
