+++ TransitionPaths/FHorz/subcodes/ValueFnSingleStep/QuasiHyperbolic/fastOLG/GridInterpLayer/ValueFnIter_FHorz_TPath_SingleStep_QHS_fastOLG_GI1_nod_noz_raw.m
@@ -60,7 +60,7 @@ inUpperStrict = (maxindexL2 >= n2short+3) & (maxindexL2 <= n2long-1);
 Policy(3,:,:) = shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
 %% Vunderbar: re-evaluate at Policy's aprime with beta
-linidx=double(reshape(maxindexL2,[1,N_a*N_j]))+n2long*(0:N_a*N_j-1);
+linidx=reshape(maxindexL2,[1,N_a*N_j])+n2long*(0:N_a*N_j-1);
 EV_at_policy=reshape(EVfine(linidx),[N_a,N_j]);
 V=shiftdim(Vhatii,1)+reshape(beta_J-beta0beta_J,[1,N_j]).*EV_at_policy;
 Vhat=shiftdim(Vhatii,1); % snapshot pre-Vunderbar
