+++ TransitionPaths/FHorz/subcodes/ValueFnSingleStep/QuasiHyperbolic/fastOLG/GridInterpLayer/ValueFnIter_FHorz_TPath_SingleStep_QHS_fastOLG_GI1_raw.m
@@ -94,7 +94,7 @@ if vfoptions.lowmemory==0
     Policy(4,:,:,:)=shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
     %% Vunderbar: re-evaluate at Policy's aprime with beta
-    linidx=double(reshape(maxindexL2,[1,N_a*N_j*N_z]))+N_d*n2long*(0:N_a*N_j*N_z-1);
+    linidx=reshape(maxindexL2,[1,N_a*N_j*N_z])+N_d*n2long*(0:N_a*N_j*N_z-1);
     EV_at_policy=reshape(EVfine(linidx),[N_a,N_j,N_z]);
     V=shiftdim(Vhatii,1)+reshape(beta_J-beta0beta_J,[1,N_j,1]).*EV_at_policy;
     V=reshape(V,[N_a*N_j,N_z]);
@@ -139,7 +139,7 @@ elseif vfoptions.lowmemory==1
         Policy(4,:,:,z_c)=shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
         %% Vunderbar
-        linidx_z=double(reshape(maxindexL2,[1,N_a*N_j]))+N_d*n2long*(0:N_a*N_j-1);
+        linidx_z=reshape(maxindexL2,[1,N_a*N_j])+N_d*n2long*(0:N_a*N_j-1);
         EV_at_policy_z=reshape(EVfine_z(linidx_z),[N_a,N_j]);
         Vtemp=shiftdim(Vhatii,1)+reshape(beta_J-beta0beta_J,[1,N_j]).*EV_at_policy_z;
         V(:,z_c)=reshape(Vtemp,[N_a*N_j,1]);
