+++ TransitionPaths/FHorz/subcodes/ValueFnSingleStep/QuasiHyperbolic/slowOLG/DivideConquer/GridInterpLayer/ValueFnIter_FHorz_TPath_SingleStep_QHS_DC1_GI1_nod_noz_raw.m
@@ -94,7 +94,7 @@ for reverse_j=1:N_j-1
     inUpperStrict = (maxindexL2 >= n2short+3) & (maxindexL2 <= n2long-1);
     Policy(3,:,jj) = shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
-    linidx=double(reshape(maxindexL2,[1,N_a]))+n2long*(0:N_a-1);
+    linidx=reshape(maxindexL2,[1,N_a])+n2long*(0:N_a-1);
     EV_at_policy=reshape(EVfine(linidx),[N_a,1]);
     V(:,jj)=Vhat_jj+(beta-beta0beta)*EV_at_policy;
 end
